requirements.txt
Flask==2.2.3
google-api-python-client==2.70.0
google-auth==2.16.2
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
pandas==1.5.3
numpy==1.24.2
python-dotenv==1.0.0
gunicorn==20.1.0
Werkzeug==2.2.3
Jinja2==3.1.2
itsdangerous==2.1.2
MarkupSafe==2.1.2
requests==2.28.2 